requirements.txt
aiohttp==3.9.5
lxml==5.2.2
python-dotenv==1.0.1
orjson==3.10.7
backoff==2.2.1
